requirements.txt
Django
django-cors-headers
django-extensions
django-nested-admin
django-phonenumber-field
django-soft-delete
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg
drfasyncview
gunicorn
mysqlclient
phonenumberslite
PyJWT
pytest
python-dateutil
uvicorn
